+++ opentelemetry-prometheus/src/lib.rs
@@ -93,7 +93,7 @@
 )]
 #![cfg_attr(test, deny(warnings))]
 
-use once_cell::sync::OnceCell;
+use once_cell::sync::{Lazy, OnceCell};
 use opentelemetry_api::{
     global,
     metrics::{MetricsError, Result, Unit},
@@ -112,6 +112,7 @@ use prometheus::{
     proto::{LabelPair, MetricFamily, MetricType},
 };
 use std::{
+    any::TypeId,
     borrow::Cow,
     collections::{BTreeMap, HashMap},
     sync::{Arc, Mutex},
@@ -126,8 +127,8 @@ const SCOPE_INFO_DESCRIPTION: &str = "Instrumentation Scope metadata";
 
 const SCOPE_INFO_KEYS: [&str; 2] = ["otel_scope_name", "otel_scope_version"];
 
-// prometheus counters MUST have a _total suffix:
-// https://github.com/open-telemetry/opentelemetry-specification/blob/v1.19.0/specification/compatibility/prometheus_and_openmetrics.md?plain=1#L282
+// prometheus counters MUST have a _total suffix by default:
+// https://github.com/open-telemetry/opentelemetry-specification/blob/v1.20.0/specification/compatibility/prometheus_and_openmetrics.md
 const COUNTER_SUFFIX: &str = "_total";
 
 mod config;
@@ -185,6 +186,7 @@ struct Collector {
     reader: Arc<ManualReader>,
     disable_target_info: bool,
     without_units: bool,
+    without_counter_suffixes: bool,
     disable_scope_info: bool,
     create_target_info_once: OnceCell<MetricFamily>,
     namespace: Option<String>,
@@ -197,7 +199,65 @@ struct CollectorInner {
     metric_families: HashMap<String, MetricFamily>,
 }
 
+// TODO: Remove lazy and switch to pattern matching once `TypeId` is stable in
+// const context: https://github.com/rust-lang/rust/issues/77125
+static HISTOGRAM_TYPES: Lazy<[TypeId; 3]> = Lazy::new(|| {
+    [
+        TypeId::of::<data::Histogram<i64>>(),
+        TypeId::of::<data::Histogram<u64>>(),
+        TypeId::of::<data::Histogram<f64>>(),
+    ]
+});
+static SUM_TYPES: Lazy<[TypeId; 3]> = Lazy::new(|| {
+    [
+        TypeId::of::<data::Sum<i64>>(),
+        TypeId::of::<data::Sum<u64>>(),
+        TypeId::of::<data::Sum<f64>>(),
+    ]
+});
+static GAUGE_TYPES: Lazy<[TypeId; 3]> = Lazy::new(|| {
+    [
+        TypeId::of::<data::Gauge<i64>>(),
+        TypeId::of::<data::Gauge<u64>>(),
+        TypeId::of::<data::Gauge<f64>>(),
+    ]
+});
+
 impl Collector {
+    fn metric_type_and_name(&self, m: &data::Metric) -> Option<(MetricType, Cow<'static, str>)> {
+        let mut name = self.get_name(m);
+
+        let data = m.data.as_any();
+        let type_id = data.type_id();
+
+        if HISTOGRAM_TYPES.contains(&type_id) {
+            Some((MetricType::HISTOGRAM, name))
+        } else if GAUGE_TYPES.contains(&type_id) {
+            Some((MetricType::GAUGE, name))
+        } else if SUM_TYPES.contains(&type_id) {
+            let is_monotonic = if let Some(v) = data.downcast_ref::<data::Sum<i64>>() {
+                v.is_monotonic
+            } else if let Some(v) = data.downcast_ref::<data::Sum<u64>>() {
+                v.is_monotonic
+            } else if let Some(v) = data.downcast_ref::<data::Sum<f64>>() {
+                v.is_monotonic
+            } else {
+                false
+            };
+
+            if is_monotonic {
+                if !self.without_counter_suffixes {
+                    name = format!("{name}{COUNTER_SUFFIX}").into();
+                }
+                Some((MetricType::COUNTER, name))
+            } else {
+                Some((MetricType::GAUGE, name))
+            }
+        } else {
+            None
+        }
+    }
+
     fn get_name(&self, m: &data::Metric) -> Cow<'static, str> {
         let name = sanitize_name(&m.name);
         match (
@@ -271,28 +331,38 @@ impl prometheus::core::Collector for Collector {
             };
 
             for metrics in scope_metrics.metrics {
-                let name = self.get_name(&metrics);
-                let description = metrics.description;
-                let data = metrics.data.as_any();
+                let (metric_type, name) = match self.metric_type_and_name(&metrics) {
+                    Some((metric_type, name)) => (metric_type, name),
+                    _ => continue,
+                };
+
                 let mfs = &mut inner.metric_families;
+                let (drop, help) = validate_metrics(&name, &metrics.description, metric_type, mfs);
+                if drop {
+                    continue;
+                }
+
+                let description = help.unwrap_or_else(|| metrics.description.into());
+                let data = metrics.data.as_any();
+
                 if let Some(hist) = data.downcast_ref::<data::Histogram<i64>>() {
-                    add_histogram_metric(&mut res, hist, description, &scope_labels, name, mfs);
+                    add_histogram_metric(&mut res, hist, description, &scope_labels, name);
                 } else if let Some(hist) = data.downcast_ref::<data::Histogram<u64>>() {
-                    add_histogram_metric(&mut res, hist, description, &scope_labels, name, mfs);
+                    add_histogram_metric(&mut res, hist, description, &scope_labels, name);
                 } else if let Some(hist) = data.downcast_ref::<data::Histogram<f64>>() {
-                    add_histogram_metric(&mut res, hist, description, &scope_labels, name, mfs);
+                    add_histogram_metric(&mut res, hist, description, &scope_labels, name);
                 } else if let Some(sum) = data.downcast_ref::<data::Sum<u64>>() {
-                    add_sum_metric(&mut res, sum, description, &scope_labels, name, mfs);
+                    add_sum_metric(&mut res, sum, description, &scope_labels, name);
                 } else if let Some(sum) = data.downcast_ref::<data::Sum<i64>>() {
-                    add_sum_metric(&mut res, sum, description, &scope_labels, name, mfs);
+                    add_sum_metric(&mut res, sum, description, &scope_labels, name);
                 } else if let Some(sum) = data.downcast_ref::<data::Sum<f64>>() {
-                    add_sum_metric(&mut res, sum, description, &scope_labels, name, mfs);
+                    add_sum_metric(&mut res, sum, description, &scope_labels, name);
                 } else if let Some(g) = data.downcast_ref::<data::Gauge<u64>>() {
-                    add_gauge_metric(&mut res, g, description, &scope_labels, name, mfs);
+                    add_gauge_metric(&mut res, g, description, &scope_labels, name);
                 } else if let Some(g) = data.downcast_ref::<data::Gauge<i64>>() {
-                    add_gauge_metric(&mut res, g, description, &scope_labels, name, mfs);
+                    add_gauge_metric(&mut res, g, description, &scope_labels, name);
                 } else if let Some(g) = data.downcast_ref::<data::Gauge<f64>>() {
-                    add_gauge_metric(&mut res, g, description, &scope_labels, name, mfs);
+                    add_gauge_metric(&mut res, g, description, &scope_labels, name);
                 }
             }
         }
@@ -363,20 +433,12 @@ fn validate_metrics(
 fn add_histogram_metric<T: Numeric>(
     res: &mut Vec<MetricFamily>,
     histogram: &data::Histogram<T>,
-    mut description: Cow<'static, str>,
+    description: String,
     extra: &[LabelPair],
     name: Cow<'static, str>,
-    mfs: &mut HashMap<String, MetricFamily>,
 ) {
     // Consider supporting exemplars when `prometheus` crate has the feature
     // See: https://github.com/tikv/rust-prometheus/issues/393
-    let (drop, help) = validate_metrics(&name, &description, MetricType::HISTOGRAM, mfs);
-    if drop {
-        return;
-    }
-    if let Some(help) = help {
-        description = help.into();
-    }
 
     for dp in &histogram.data_points {
         let kvs = get_attrs(&mut dp.attributes.iter(), extra);
@@ -404,7 +466,7 @@ fn add_histogram_metric<T: Numeric>(
 
         let mut mf = prometheus::proto::MetricFamily::default();
         mf.set_name(name.to_string());
-        mf.set_help(description.to_string());
+        mf.set_help(description.clone());
         mf.set_field_type(prometheus::proto::MetricType::HISTOGRAM);
         mf.set_metric(protobuf::RepeatedField::from_vec(vec![pm]));
         res.push(mf);
@@ -414,26 +476,15 @@ fn add_histogram_metric<T: Numeric>(
 fn add_sum_metric<T: Numeric>(
     res: &mut Vec<MetricFamily>,
     sum: &data::Sum<T>,
-    mut description: Cow<'static, str>,
+    description: String,
     extra: &[LabelPair],
-    mut name: Cow<'static, str>,
-    mfs: &mut HashMap<String, MetricFamily>,
+    name: Cow<'static, str>,
 ) {
-    let metric_type;
-    if sum.is_monotonic {
-        name = format!("{name}{COUNTER_SUFFIX}").into();
-        metric_type = MetricType::COUNTER;
+    let metric_type = if sum.is_monotonic {
+        MetricType::COUNTER
     } else {
-        metric_type = MetricType::GAUGE;
-    }
-
-    let (drop, help) = validate_metrics(&name, &description, metric_type, mfs);
-    if drop {
-        return;
-    }
-    if let Some(help) = help {
-        description = help.into();
-    }
+        MetricType::GAUGE
+    };
 
     for dp in &sum.data_points {
         let kvs = get_attrs(&mut dp.attributes.iter(), extra);
@@ -453,7 +504,7 @@ fn add_sum_metric<T: Numeric>(
 
         let mut mf = prometheus::proto::MetricFamily::default();
         mf.set_name(name.to_string());
-        mf.set_help(description.to_string());
+        mf.set_help(description.clone());
         mf.set_field_type(metric_type);
         mf.set_metric(protobuf::RepeatedField::from_vec(vec![pm]));
         res.push(mf);
@@ -463,19 +514,10 @@ fn add_sum_metric<T: Numeric>(
 fn add_gauge_metric<T: Numeric>(
     res: &mut Vec<MetricFamily>,
     gauge: &data::Gauge<T>,
-    mut description: Cow<'static, str>,
+    description: String,
     extra: &[LabelPair],
     name: Cow<'static, str>,
-    mfs: &mut HashMap<String, MetricFamily>,
 ) {
-    let (drop, help) = validate_metrics(&name, &description, MetricType::GAUGE, mfs);
-    if drop {
-        return;
-    }
-    if let Some(help) = help {
-        description = help.into();
-    }
-
     for dp in &gauge.data_points {
         let kvs = get_attrs(&mut dp.attributes.iter(), extra);